backend/requirements.txt
Flask
mysqlclient
python-dotenv
os
importlib
json